Check and handle failed jobs
